refactor(theme): move theme to a separate folder
