+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from './components'
-export * from './contexts/theme/provider'
+export * from './theme/provider'
